Adding the deps to the external package for the apis
